feat(createPath): createPage connects to endpoint
