+++ src/frontend/containers/CreatePage.tsx
@@ -1,51 +1,60 @@
-// import React, { FC, useState } from 'react';
+import React, { FC, useState } from 'react';
+// import { Path } from '../../shared/models/Path';
+import { Button, Form, Input, Label } from 'reactstrap';
 
-// interface FormElements extends HTMLFormElement {
-//   nameInput: HTMLInputElement;
-// }
+export const CreatePage: FC = () => {
+  const [path, setPath] = useState({
+    name: '',
+    description: '',
+  });
 
-// interface CreateFormElements extends HTMLFormElement {
-//   readonly elements: FormElements;
-// }
+  //when name updates
+  const nameChange = (name: string) => {
+    setPath({ ...path, name: name });
+  };
 
-// export const CreatePage: FC = () => {
-//   const { path, setPath } = useState<object>({
-//     id: '',
-//     name: '',
-//     description: '',
-//     owner: '',
-//     waypoints: [],
-//   });
+  //when description updates
+  const descriptionChange = (description: string) => {
+    setPath({ ...path, description: description });
+  };
 
-//   // const handleSubmit = useCallback(
-//   //   (pathDetails: object) => {
-//   //     console.log(pathDetails);
-//   //     setPath(pathDetails);
-//   //   },
-//   //   [path],
-//   // );
+  //creates new path and route to new page
+  const clickCreate = async () => {
+    await api.createPath(path.name, path.description); //call to create path endpoint
+    //route to edit page
+  };
 
-//   const handleSubmit = (event: React.FormEvent<CreateFormElements>) => {
-//     console.log(event.currentTarget.elements);
-//   };
+  return (
+    <>
+      <h1>Create a new Path</h1>
 
-//   // Title
-//   // Description
-//   // Display current waypoints
-//   // Display input for a new waypoint
-//   return (
-//     <>
-//       <h1>Create a new Path</h1>
-
-//       <form onSubmit={handleSubmit(onsubmit())}>
-//         <label htmlFor={'nameInput'}>Name</label>
-//         <input type={'text'} id={'nameInput'}></input>
-//         <label htmlFor={'descriptionInput'}>Description</label>
-//         <input type={'text'} id={'descriptionInput'}></input>
-//         <button type={'submit'}>
-//           Create
-//         </button>
-//       </form>
-//     </>
-//   );
-// };
+      <Form>
+        <Label for={'nameInput'}>Name</Label>
+        <Input
+          type={'text'}
+          id={'nameInput'}
+          onChange={(e) => {
+            nameChange(e.currentTarget.value); //when name is updated
+          }}
+        ></Input>
+        <Label for={'descriptionInput'}>Description</Label>
+        <Input
+          type={'text'}
+          id={'descriptionInput'}
+          onChange={(e) => {
+            descriptionChange(e.currentTarget.value); //when description is updated
+          }}
+        ></Input>
+        <Button
+          type={'submit'}
+          id={'createButton'}
+          onClick={() => async () => {
+            await clickCreate();
+          }}
+        >
+          Create
+        </Button>
+      </Form>
+    </>
+  );
+};
